requirements.txt
numpy>=1.24
Pillow>=9.0
PyYAML>=6.0
torch>=2.1
torchvision>=0.16
tqdm>=4.66
